Initial kickoff
This commit adds a first idea how to structure the code.
The implementations in this project will be used in
services provided by the CSPs. We will start with AWS,
others might follow. The python implementations are
pyproject and poetry based and uses their own python
namespaces such that we can package and version them
individually while still on the same source git to
follow all changes. The commit from here adds a very
rough skeleton for two services in AWS. There is
the resolve_customer namespace for adding code that
uses the AWS ResolveCustomer API and there is the
sqs_event_manager which is expected to add code that
runs as AWS lambda triggered by SQS events. As delivery
method we will use OCI container(s) for which the
container path was added and kiwi image descriptions
are expected to follow. To build packages from the
individual python components/namespaces I have added
a concept that consistently build packages sources
from a template plus poetry and a make package
Makefile target. There are still some parts to discuss,
and some open questions.