Added resolve_customer lambda handler
The resolve_customer handler expects an event providing
the x-amzn-marketplace-token. It will then resolve customer
information from the metering marketplace as well as the
entitlements for this customer's ID and product code.
The input and return API is documented in the header of the
lambda_handler entrypoint and follows the AWS API Gateway
specifications.