Add new class AWSAssumeRole
An instance of AWSAssumeRole allows to fetch temporary
credentials to access API endpoints from an assumed role
arn that can point to another AWS account and IAM role.
We make use of this concept because all marketplace
accounts will live outside of the account that setups
the SUSE SaaS tool chain. The required configuration data
to construct an instance of AWSAssumeRole are taken from
a configuration file named /etc/assume_role.yml. The
configuration file is packaged as a sub-package and can
be provided with different content in a production
setup.