Various fixes to the sqs event manager
I did a setup of an sqs queue in AWS and connected it to
the sqs event manager as a lambda. Next I sent an event to the
queue that should simulate an entitlement-updated event as
follows:

```
session = boto3.Session(profile_name='suse-pct')
sqs = session.resource('sqs')
alias_queue sqs.get_queue_by_name(QueueName='ms-testing.fifo')
data = json.dumps(
    {
        'Type' : 'Notification',
        'MessageId' : 'uuid',
        'TopicArn' : 'arn:aws:sns:us-east-1:XXX:aws-mp-entitlement-notification-XXX',
        'Message' : {
            'action': 'entitlement-updated',
            'customer-identifier': 'string',
            'product-code': 'string'
        },
        'Timestamp' : '2025-01-15 16:31:50',
        'SignatureVersion' : '1',
        'Signature' : 'abc123',
        'SigningCertURL' : 'string',
        'UnsubscribeURL' : 'string'
    }
)
alias_queue.send_message(
    MessageBody=data, MessageGroupId='123456789'
)
```

I receieved errors on cloud watch for the code running in the
lambda and was debugging them. The result is this PR which
changes the following bits:

* print event and context to the log for better debugging
* allow the message data to be processed in different
  quotation annotations
* Cleanup and fixed the unit tests and typing annotations
* Activate github actions for sqs event manager code
* Use 1:1 copy of the event data for unit testing, deleted
  sensitive information from the record
* Cleanup the logging