Fixed marketplace polling
Only a successful assume_role() is not enough to know if
the subsequent marketplace API can be reached with that
role credentials. Thus perform the marketplace call with
that credentials and loop to the next role if it has
failed. Also report all errors in case none of the attempts
succeeds