Fix product code retrieval
The product code was taken from the first entitlement
list entry. This is wrong because there could be a
product code without any entitlements. Thus the product
code has to be taken from the response of the resolve
customer API and not from the response of the entitlements.
This Fixes #28