Improve error reporting for sqs event manager
If an event appears we are creating a POST request against
a configured endpoint. This commit includes the HTTP status
code to the error message in case of an error (!= 200). In
addition the reported response text depends on the endpoint
implementation. It can happen that the endpoint does not
include any message, in this case we replace the information
with a "no response text" provided such that we know the
endpoint has an issue but no further information was provided.