Improve handling of other event categories
All Message data that we handle have the general message Type
category set to 'Notification'. There are other categories
like 'SubscriptionNotification' which are events sent when
someone subscribes a new SNS topic. Probably there are more
Type categories but we want to handle only the ones that comes
in as 'Notification'.